Adding a bootstrap and core source configuration.
In order for us to use Hiera effectively in our server management
we must have Hiera installed and configured.  But Hiera is used to
install and configure software.

In order to deal with this situation we create a separate bootstrap
process but leave most of the installation to the specialized
server configuration.